보드게시판 CRUD 기능 구현

+++ src/main/java/com/kh/pp/auth/model/vo/CustomUserDetails.java
@@ -11,7 +11,7 @@ import lombok.Value;
 @Value
 @Builder
 public class CustomUserDetails implements UserDetails {
-	private int memberNo;
+	private Long memberNo;
 	
 	private String username; // = Admin_Id 
 	private String password;
